test(textarea): migrate fill
