+++ core/src/components/textarea/test/fill/textarea.e2e.ts-snapshots/textarea-fill-outline-label-floating-md-ltr-Mobile-Safari-linux.png
