+++ core/src/components/textarea/test/fill/textarea.e2e.ts-snapshots/textarea-fill-outline-label-floating-md-rtl-Mobile-Chrome-linux.png
