+++ core/src/components/textarea/test/fill/textarea.e2e.ts-snapshots/textarea-fill-shaped-outline-custom-md-rtl-Mobile-Chrome-linux.png
