+++ core/src/components/textarea/test/fill/textarea.e2e.ts-snapshots/textarea-fill-shaped-outline-md-ltr-Mobile-Chrome-linux.png
