+++ core/src/components/textarea/test/fill/textarea.e2e.ts-snapshots/textarea-fill-shaped-solid-custom-md-rtl-Mobile-Firefox-linux.png
