+++ core/src/components/textarea/test/fill/textarea.e2e.ts-snapshots/textarea-fill-shaped-solid-md-ltr-Mobile-Firefox-linux.png
