+++ core/src/components/textarea/test/fill/textarea.e2e.ts-snapshots/textarea-fill-shaped-solid-md-rtl-Mobile-Chrome-linux.png
