+++ core/src/components/textarea/test/fill/textarea.e2e.ts-snapshots/textarea-fill-solid-label-floating-md-rtl-Mobile-Safari-linux.png
